+++ src/main/java/org/jabref/gui/entryeditor/LatexCitationsTabViewModel.java
@@ -15,10 +15,16 @@ import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 
 import org.jabref.gui.AbstractViewModel;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.preferences.GuiPreferences;
+import org.jabref.gui.push.PushToApplication;
+import org.jabref.gui.push.PushToApplications;
+import org.jabref.gui.push.PushToTeXstudio;
+import org.jabref.gui.texparser.CitationsDisplay;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.gui.util.UiTaskExecutor;
 import org.jabref.logic.l10n.Localization;
@@ -152,6 +158,22 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
             public void onDirectoryDelete(File directory) {
             }
         };
+    }
+
+    public void handleMouseClick(MouseEvent event, CitationsDisplay citationsDisplay) {
+        Citation selectedItem = citationsDisplay.getSelectionModel().getSelectedItem();
+
+        if (event.getButton() == MouseButton.PRIMARY && event.getClickCount() == 2 && selectedItem != null) {
+            String applicationName = preferences.getPushToApplicationPreferences()
+                                                .getActiveApplicationName();
+            PushToApplication application = PushToApplications.getApplicationByName(
+                                                                      applicationName,
+                                                                      dialogService,
+                                                                      preferences)
+                                                              .orElse(new PushToTeXstudio(dialogService, preferences));
+            preferences.getPushToApplicationPreferences().setActiveApplicationName(application.getDisplayName());
+            application.jumpToLine(selectedItem.path(), selectedItem.line(), selectedItem.colStart());
+        }
     }
 
     public void bindToEntry(BibEntry entry) {